# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Fiber Merchant Kit API",
    "version": "1.0.0",
    "description": "Stripe-style merchant API for Fiber Network invoices, webhooks, transactions, balances, and settlement operations."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local API server"
    }
  ],
...
  "paths": {
...
    "/api/v1/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get authenticated merchant context",
        "operationId": "getAuthContext",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated merchant context and permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantAuthContext"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/auth/api-key/rotate": {
      "post": {
        "tags": ["Auth"],
        "summary": "Rotate the current merchant API key",
        "operationId": "rotateApiKey",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Replacement API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyRotationResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/v1/transactions": {
      "get": {
        "tags": ["Transactions"],
        "summary": "List transactions",
        "operationId": "listTransactions",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/TransactionStatus"
            }
          },
          {
            "name": "direction",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["incoming", "outgoing"]
            }
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated transactions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionList"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/transactions/{id}": {
      "get": {
        "tags": ["Transactions"],
        "summary": "Get a transaction",
        "operationId": "getTransaction",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/balance/channels": {
      "get": {
        "tags": ["Balances"],
        "summary": "List Fiber channel balances",
        "operationId": "listChannelBalances",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Channel balances",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChannelBalance"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/balance/total": {
      "get": {
        "tags": ["Balances"],
        "summary": "Get total Fiber channel balance",
        "operationId": "getTotalBalance",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Total balance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TotalBalance"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/stats": {
      "get": {
        "tags": ["Stats"],
        "summary": "Get dashboard stats",
        "operationId": "getStats",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard stats",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Stats"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/stats/revenue": {
      "get": {
        "tags": ["Stats"],
        "summary": "Get revenue history",
        "operationId": "getRevenueHistory",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 365,
              "default": 30
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Revenue history rows",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RevenueRow"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/fiber/status": {
      "get": {
        "tags": ["Fiber"],
        "summary": "Get Fiber node and worker status",
        "operationId": "getFiberStatus",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Fiber node, channel, and settlement worker status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FiberStatus"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/fiber/settlement/run": {
      "post": {
        "tags": ["Fiber"],
        "summary": "Run settlement sweep",
        "operationId": "runSettlementSweep",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Settlement sweep completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettlementRun"
                }
              }
            }
          },
          "202": {
            "description": "Settlement sweep skipped because another run is active",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettlementRun"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "fm_sk"
      }
    },
    "parameters": {
      "Id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 50
        }
      },
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "schema": {
          "type": "string"
        },
        "description": "Opaque pagination cursor returned by the previous page."
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid API key",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated API key does not have the required role",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "ValidationError": {
        "description": "Validation failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "MerchantRole": {
        "type": "string",
        "enum": ["owner", "admin", "developer", "viewer"]
      },
      "MerchantUser": {
        "type": "object",
        "required": ["id", "merchantId", "email", "role", "active", "createdAt"],
        "properties": {
          "id": {
            "type": "string"
          },
          "merchantId": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "$ref": "#/components/schemas/MerchantRole"
          },
          "active": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MerchantAuthContext": {
        "type": "object",
        "required": ["merchantId", "role", "permissions", "users"],
        "properties": {
          "merchantId": {
            "type": "string"
          },
          "label": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "$ref": "#/components/schemas/MerchantRole"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MerchantUser"
            }
          }
        }
      },
      "ApiKeyRotationResponse": {
        "type": "object",
        "required": ["merchantId", "apiKey", "role", "rotatedAt"],
        "properties": {
          "merchantId": {
            "type": "string"
          },
          "apiKey": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/MerchantRole"
          },
          "rotatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Transaction": {
        "type": "object",
        "required": ["id", "paymentHash", "direction", "amount", "currency", "status", "createdAt"],
        "properties": {
          "id": {
            "type": "string"
          },
          "paymentHash": {
            "type": "string"
          },
          "invoiceId": {
            "type": "string",
            "nullable": true
          },
          "direction": {
            "type": "string",
            "enum": ["incoming", "outgoing"]
          },
          "amount": {
            "type": "string"
          },
          "currency": {
            "$ref": "#/components/schemas/Currency"
          },
          "fee": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/TransactionStatus"
          },
          "counterparty": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "metadata": {
            "type": "object",
            "nullable": true,
            "additionalProperties": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TransactionList": {
        "type": "object",
        "required": ["items", "total"],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Transaction"
            }
          },
          "total": {
            "type": "integer"
          },
          "cursor": {
            "type": "string"
          }
        }
      },
      "ChannelBalance": {
        "type": "object",
        "required": ["localBalance", "remoteBalance", "capacity", "asset"],
        "properties": {
          "localBalance": {
            "type": "string"
          },
          "remoteBalance": {
            "type": "string"
          },
          "capacity": {
            "type": "string"
          },
          "asset": {
            "type": "string"
          },
          "channelId": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "peerPubkey": {
            "type": "string"
          }
        }
      },
      "TotalBalance": {
        "type": "object",
        "required": ["local", "remote", "total"],
        "properties": {
          "local": {
            "type": "string"
          },
          "remote": {
            "type": "string"
          },
          "total": {
            "type": "string"
          }
        }
      },
      "Stats": {
        "type": "object",
        "required": ["totalInvoices", "paidInvoices", "totalVolume", "successRate", "activeChannels", "channelBalances"],
        "properties": {
          "totalInvoices": {
            "type": "integer"
          },
          "paidInvoices": {
            "type": "integer"
          },
          "totalVolume": {
            "type": "string"
          },
          "successRate": {
            "type": "integer"
          },
          "activeChannels": {
            "type": "integer"
          },
          "channelBalances": {
            "type": "object",
            "required": ["local", "remote"],
            "properties": {
              "local": {
                "type": "string"
              },
              "remote": {
                "type": "string"
              }
            }
          }
        }
      },
      "RevenueRow": {
        "type": "object",
        "required": ["date", "volume", "count"],
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "volume": {
            "type": "number"
          },
          "count": {
            "type": "integer"
          }
        }
      },
      "FiberStatus": {
        "type": "object",
        "required": ["mode", "reachable", "rpcUrlConfigured", "rpcEndpoints", "node", "channels", "worker"],
        "properties": {
          "mode": {
            "type": "string",
            "enum": ["demo", "live"]
          },
          "reachable": {
            "type": "boolean"
          },
          "rpcUrlConfigured": {
            "type": "boolean"
          },
          "rpcEndpoints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FiberRpcEndpointStatus"
            }
          },
          "node": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/FiberNodeStatus"
              }
            ]
          },
          "channels": {
            "$ref": "#/components/schemas/FiberChannelStatus"
          },
          "worker": {
            "$ref": "#/components/schemas/SettlementWorkerStatus"
          },
          "error": {
            "type": "string"
          }
        }
      },
      "FiberRpcEndpointStatus": {
        "type": "object",
        "required": ["url", "reachable"],
        "properties": {
          "url": {
            "type": "string"
          },
          "reachable": {
            "type": "boolean"
          },
          "nodeId": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        }
      },
      "FiberNodeStatus": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "nodeId": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "peersCount": {
            "type": "integer"
          },
          "channelsCount": {
            "type": "integer"
          },
          "pendingChannelsCount": {
            "type": "integer"
          }
        }
      },
      "FiberChannelStatus": {
        "type": "object",
        "required": ["total", "ready", "pending", "failed", "localBalance", "remoteBalance", "totalCapacity", "items"],
        "properties": {
          "total": {
            "type": "integer"
          },
          "ready": {
            "type": "integer"
          },
          "pending": {
            "type": "integer"
          },
          "failed": {
            "type": "integer"
          },
          "localBalance": {
            "type": "string"
          },
          "remoteBalance": {
            "type": "string"
          },
          "totalCapacity": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChannelBalance"
            }
          }
        }
      },
      "SettlementWorkerStatus": {
        "type": "object",
        "required": ["enabled", "active", "running", "mode", "intervalMs", "batchSize"],
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "active": {
            "type": "boolean"
          },
          "running": {
            "type": "boolean"
          },
          "mode": {
            "type": "string",
            "enum": ["demo", "live"]
          },
          "intervalMs": {
            "type": "integer"
          },
          "batchSize": {
            "type": "integer"
          },
          "lastRunAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastSuccessAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastError": {
            "type": "string"
          },
          "lastSummary": {
            "$ref": "#/components/schemas/SettlementSummary"
          }
        }
      },
      "SettlementRun": {
        "type": "object",
        "required": ["trigger", "running", "skipped", "startedAt"],
        "properties": {
          "trigger": {
            "type": "string",
            "enum": ["timer", "manual"]
          },
          "running": {
            "type": "boolean"
          },
          "skipped": {
            "type": "boolean"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "summary": {
            "$ref": "#/components/schemas/SettlementSummary"
          },
          "error": {
            "type": "string"
          }
        }
      },
      "SettlementSummary": {
        "type": "object",
        "required": ["checked", "paid", "received", "expired", "unchanged", "errors"],
        "properties": {
          "checked": {
            "type": "integer"
          },
          "paid": {
            "type": "integer"
          },
          "received": {
            "type": "integer"
          },
          "expired": {
            "type": "integer"
          },
          "unchanged": {
            "type": "integer"
          },
          "errors": {
            "type": "integer"
          }
        }
      },
      "Currency": {
        "type": "string",
        "enum": ["CKB", "RUSD"],
        "default": "CKB"
      },
...
      "TransactionStatus": {
        "type": "string",
        "enum": ["Pending", "Succeeded", "Failed", "Timeout", "Abandoned"]
      },
...
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "required": ["error", "details"],
        "properties": {
          "error": {
            "type": "string",
            "example": "Validation failed"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    }
  }
}